refactor(edit): Group array logic
